[netstack3] Test-only Eq impl for SocketCookie
The motivation for this is for use in diagnostics tests, where we don't
want to have to take ownership over the value just to test that the
cookie is correct by exporting it.

It ended up being a somewhat viral change because it required adding a
new testutils variant of the netstack3-sync and, transitively,
netstack3-trace crates.

Change-Id: I67e38487ebded481cef6e86a9ebfe5256a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409554
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>