[blobfs] align default cache policy with fshost
fshost defaults to (and so almost universally overrides) the cache
eviction policy to never_evict, so there is no reason to have a
different default in blobfs.

This was causing a bug with the new fvm2 component, because it wasn't
forwarding the override of this option, so it was evicting by default.

This option should just ultimately be removed, but this fixes the
immediate issue.

Bug: 456807639
Change-Id: I32409f99d0d55c9753cc8dff08dd9e8a3e772780
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409897
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>