[host-target-testing] Make uncompressed blobs available if needed
After https://fxrev.dev/1351551, uncompressed blobs are not included by
default in local builds. This causes local OTA tests to fail in test
cases where a new update package is created on-the-fly.
This change ensures that uncompressed blobs for ZBI and vbmeta are
available for the test to modify as needed and create the new update
package.

Fixed: 446897798
Change-Id: I5a170d10b6d96462758b447acda1018138f9d93a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409914
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>