[lockdep] Check if reporting is disabled for AcquireAfterLeaf
If a mutex is tagged as LockFlagsLeaf and is contended, the lock
acquisition path will try to update the thread's runtime stats when
blocking. This acquires the ThreadRuntimeStats lock which is tagged with
LockFlagsReportingDisabled, but a violation is reported nevertheless
without this fix.

Run-All-Tests: true
Change-Id: I88ba286696fb26ab4087f723371fd958818e5fd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409703
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>