[kernel] Encode kernel aspace base address in ZirconAbiSpec
This change extends the ZirconAbiSpec to include the kernel's choice of
address space base, removing the duplicated constants on the phys side
and in effect making this an opaque decision to physboot.

Initial thinking was physboot could just always pick the minimal,
hardware-prescribed address. This is what we do for arm64 and riscv64,
but on x86 it was overlooked that we actually intentionally constrain
the base address to -512GiB. The reason for this was to avoid more
complicated synchronization (e.g., cross-CPU TLB shootdowns) when
linking/unlinking top-level entries in the kernel half of the root page
table. By 'burning' all but the last entry and ensuring the last entry
is never unlinked, we ensure that every address space we switch to has
an identical upper half in its root table. This is naturally not a
problem on arm64 with its split of kernel and user address spaces.
Further, this problem was avoided on riscv64 by pre-allocating the whole
second level of kernel page tables (1MiB of .bss with sv39) and ensuring
in the MMU code that they are never unlinked from the root.

Bug: 42146863
Change-Id: I3b0663ae153ee6dcf71523d15c327623726cfd60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1406093
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>