[test_manager][debug] Use the recursive option when attaching
This will cause DebugAgent to attach to all components within the test's
realm. Note that this doesn't change the behavior of GetProcessInfo
since we are still passing a specific filter to that request.

Update the interface to `report_all_backtraces` to allow callers to pass
a component URL to use as a filter to just dump a specific test
package's processes, rather than all currently running test suites.

Bug: 438253057
Bug: 303314982
Change-Id: I6a6a6964b860743ad19a236abfc28ad300d472b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1396614
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>