[assembly] Explicitly set product_bundle version strings.
Currently, PBs are assigned a version number which matches the version
of the SDK that was used to build it. This works, but it's a bit
disconnected from the version of the contained assembly artifacts.

The new `ffx product-bundle create` tooling changed this behavior to use
the version string from the product config. This is better, but in some
situations I think it would be best to explicitly set the PB version at
the target level. This also allows the versioning console to transition
from "build_api/build_info.json" -> "ffx product-bundle get-version",
since we can explicitly set the version to match the previous format.

This CL explicitly sets the version string on the build target for all
product bundles in the fuchsia tree. In the future we can make this
field required.

Bug: b/456261023
Change-Id: I4cc297a9265b8a4db53c5c757a5135243cbcf405
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1407674
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>