[kernel] Use PAGE_SIZE from <lib/page/size-asm.h> in assembly
We extend libpage with a generated header with #define-versions of the
constants defined in <lib/page/size.h>, intended for assembly usage. In
migrating the kernel's assembly over to using that we remove the last
usage of the <arch/defines.h> PAGE_SIZE #defines and so remove those
definitions too. Deferring removal to a follow-up change was also not an
option as this would result in conflicting definitions of PAGE_SIZE.

Bug: 42146863
Change-Id: Id12cdd9a7beec933a011289e299ce904d1666a21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1410037
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>