[windowed-stats] No longer accept timestamp on InspectedTimeMatrix::fold
`InspectedTimeMatrix::fold` acquires a lock on `TimeMatrix` before
folding the timed sample. To avoid the possibility of MonotonicityError,
change the API to accept the sample only so that it can generate the
timestamp after acquiring a lock.

The existing `InspectedTimeMatrix::fold` that accepts a timestamp is
renamed to `InspectedTimeMatrix::fold_at` because the event builders API
(which is currently unused) still uses it.

Bug: 456481189
Test: CQ
Change-Id: I01e53d6662887536c2e5b76fdbedcd2711cd624a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1408463
Commit-Queue: Kiet Tran <kiettran@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Kiet Tran <kiettran@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>