[rust-analyzer] Quick fix for check-on-save
With buildfile set to /dev/null, rust-analyzer will try to cd to /dev,
which causes check-on-save to fail because it's trying to run fx from
outside the Fuchsia tree.

It's not totally clear if this is the right long-term fix, but it at
least unblocks things for now.

Change-Id: I2500fc801d1322ab536d06018e818abe6a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413494
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>