[starnix][perfetto] Improve trace state handling
This handles when the TraceState of the trace engine updates
faster than the TraceObserver events are processed. This should
be extremely rare because the perfetto_consumer holds a ProlongedContext
which prevents the trace session from ending.

The panic which was happening in this situation is replaced by
cleaning up the perfetto_consumer and re-initializing it.

Bug: 457768643
Change-Id: I3467863fe4ae30b3a7809e38a68abd3108de2c32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413516
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>