[fxfs] Fix last_object_id computation after unlocking
When unlocking a store, we were setting the last object ID by using
`update_last_object_id`, but that assumed that the object ID was *after*
it had been encrypted, but that's not the case when a store is being
unlocked.

I think this could potentially have led to filesystem corruption because
I think it could reset the last object ID to an earlier value than it
should which could potentially lead to objects being overwritten when
they shouldn't.

Bug: 453445060
Bug: 450789112
Change-Id: I1d00379ca2b51822c583e3abee20e35c8a9683ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413694
Commit-Queue: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>