[bazel][fidl][zither] Add remaining Zither backends
Add Zither backends for Rust, etc. `legacy_syscall_cdecl` is the
only backend still not implemented. The Go and Go runtime backends are
not implemented because we do not believe we will need them.

The following targets fail to build. This will be addressed separately.
* //sdk/fidl/zbi:zbi_zither.rust
* //sdk/fidl/zbi:zbi_zither.rust_syscall

NO_IFTTT=This change brings the Bazel file closer to parity with its the GN file.

Bug: 454449781
Test: fx bazel build --config=fuchsia //sdk/fidl/zbi:zbi_zither.asm //sdk/fidl/zbi:zbi_zither.asm.gen.check //sdk/fidl/zbi:zbi_zither.c //sdk/fidl/zbi:zbi_zither.c.gen.check //sdk/fidl/zbi:zbi_zither.kernel //sdk/fidl/zbi:zbi_zither.kernel.gen.check //sdk/fidl/zbi:zbi_zither.rust.gen.check //sdk/fidl/zbi:zbi_zither.rust_syscall.gen.check //sdk/fidl/zbi:zbi_zither.syscall_docs //sdk/fidl/zbi:zbi_zither.zircon_ifs
Change-Id: Ibab50be8191913196b696c365505908448e478e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1410095
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>