[kernel][smmu] Introduce an ARM SMMU Iommu type.
Define a type, and a parameter structure for ARM SMMUs which can be
used with the `zx_iommu_create` syscall.

Currently, systems with SMMUs are simply creating a single DummyIommu
instance, and then creating BTIs with fake StreamIDs.  This is fine
for operating with the system's SMMUs disabled, or when it is
preconfigured to put all of the required initiators into bypass mode,
but more needs to be done in the long run.

Eventually, user-mode is going to need to obtain handles to all of the
actual SMMU instances in the system, and then create BTIs using the
actual stream IDs (as reported in the device tree) under the correct
SMMU instances.

This small patch should allow for a graceful transition to that mode.
Users can now explicitly request a handle to an SMMU instance with the
desired instance uniquely identified by the physical base address of
the specific SMMU instance.  Under the hood, we are simply going to
return an instance of a DummyIommu, meaning that the system should
continue to operate exactly as it does today, allowing the plumbing to
create the proper SMMU and BTI instances to be developed in parallel
with developing support for the SMMU hardware itself.

Bug: 42108551
Change-Id: I7d22438c623a0f662aede3051d0e2c17f7b7770f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1408247
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>