[paver][moonflower] use hardcoded inactive partition type GUIDs
Previously, the paver was using the boot_a/boot_b partition flags to
determine the inactive type GUID, but we've run into some cases where
the boot partitions themselves have mismatched GUID and flag states, in
which case we would grab the wrong inactive type and totally mess up the
GPT.

This CL uses a hardcoded list of known inactive type GUIDs instead so we
know whether to swap GUIDs or not even in the case when the boot
partition is corrupted. It's still unclear how the boot partition gets
into a bad state in the first place, but this should hopefully prevent
us from totally bricking the device when it happens.

Bug: b/444508079
Change-Id: I0213bd93de09513d99b041e7bb3c31c494be6945
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413481
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: David Pursell <dpursell@google.com>