[codecs] vp9_decoder compat w new video_ucode.bin
The new video_ucode.bin needs vp9_decoder.cc to always save the input
context after each output frame decodes successfully. This is also an
incremental step toward the stream-style input handling that we do for
h264, though we don't currently need to move the rest of the way to
that strategy for vp9.

These changes also remain compatible with the old video_ucode.bin, so
this CL doesn't need to be atomic with the cipd package version update.

Also add a test checking decode of an exynos-HW-encoded stream; this
will be enabled later for some envs once cipd deps are ready; until
the video_ucode.bin version is updated to the new one, this test will
fail.

Change-Id: Id9eae17baaa07ba6a047563e74f8b31af27f68cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1411859
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>