[antlion] Remove check for fuchsia.wlan.phyimpl capability
The fuchsia.wlan.phyimpl defined protocols are associated with the
WLAN vendor driver which is generally not reachable over ffx as it's a
driver transport protocol. Moreover, wlandevicemonitor is precisely the
component that serves an abstraction over the discovered phy devices.

  wlancfg -> wlandevicemonitor -> wlanphy -> WLAN vendor driver

Since each phy device may not be immediately available after boot, it
doesn't make sense for a test to try to discover if the phy exists at
all. Instead, tests should use wlandevicemonitor to determine the
availability of phy devices, and potentially wait until they are
present.

Test: Ran start_stop_client_connections_test.sh via test-remote - http://ci.chromium.org/b/8698942820741144897
Bug: 359342196
Fixed: 452085360
Change-Id: I30adcde583dbceb89834be517173ef6f34e617f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1415036
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>