[kernel][vm] ZeroPagesLocked: Use markers more often
In ZeroPagesLocked, we sometimes have to choose between representing
zero with a freshly allocated zero page and with a marker. This CL
represents zero with markers whenever possible.

The old policy was to try to "justify" markers by pointing to committed
pages. This was believed to bound the total number of markers allocated.
We measured the number of new un-"justified" markers, and found that
this CL only increases the number of markers in rare scenarios.

Fixed: 458059877
Bug: 423017076
Change-Id: Ic4fa179e2f26bfccc4df68bfeffb0cdb2d2d5f38
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1411519
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Ethan Williams <ethanws@google.com>