[Netstack3] Keep sharing domain tokens for the lifetime of the socket
Previously SetReusePort2 was taking token koid and dropping the token
after that. Now it will keep the token for the lifetime of the socket
(or until the token is replaced). This will allow Starnix to observe
if the token is still being used.

Bug: 451615802
Change-Id: Iae3f169ab2442a2f868238b346c362475f786996
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413439
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Sergey Ulanov <sergeyu@google.com>
Fuchsia-Auto-Submit: Sergey Ulanov <sergeyu@google.com>