[build] Restore use of zircon_optimize as separate optimization control for the kernel
Through various refactorings, zircon_optimize was no longer being used
to separately set the optimization level, as the global optimize value
was being used instead.  This makes it an override, so that if it's
set, it takes precedence over the use the global optimize value.

Change-Id: I6b9c562d83dc46810339e5a212debf5352a04776
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1415582
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>