[thermal] Record max temperatures.
Records max temperatures at the same frequency that we record
historical temperatures.

Historical readings are changed to be averages, so there are natural
ways to reason about these in relation to the maxima. (Conversely, if
we were continuing to record a sample, we'd have to choose when to
record each sample in relation to the intervals on which maxima are
computed, and there isn't a natural way to do that.)

Fixed: 451015386
Change-Id: Ic8c2091abac9e77fb801e80a8ecc0e5258408589
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1391615
Reviewed-by: Michael Brunson <mbrunson@google.com>
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Commit-Queue: Onath Dillinger <claridge@google.com>