[build][api] Add should_file_changes_trigger_build() function.
This new function will be used to implement a new command that
determines whether a build is necessary after files have been
changed due to a commit being applied to the source directory.

This is bound to replace the implementation of
//tools/integration/fint/should_build.go which is based
on `gn analyze` and thus cannot see any inputs added to
the Ninja graph through depfiles. In practice it means that
the current implementation cannot see Bazel build files
and input files at all, while this new function can.

Change-Id: Ib24d40376fda5843c581d26527c129995e7408e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1411716
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>