[ffx][forward] Recover from disconnects
Introduce a sentinel listening socket on the target so we can always be
sure to detect disconnects.

The main loop watches that sentinel and triggers a delayed reconnect.
The reconnect can take a bit, we're relying entirely on the internal
timeouts from the Connector tooling here.

Adds a new argument `--once` to the CLI to opt out of the auto-reconnect
behavior, causing it to bail on the first error.

Fixed: 456723550

Change-Id: I654f3d0d26d9491c09ea86da1b70c0c66a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1415775
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>