[debugger] Improve zero line LineTable entries
We now do something closer to what GDB does:

Instead of returning one singular line table row entry for an address,
we find two entries: one for the "symbolizable" address (which doesn't
necessarily have to be is_stmt) and a "statement" address, which is the
closest following is_stmt entry after the address we're looking up.

This lets the code still be shared for breakpoint and symbolization code
with a new option to ResolveOptions to specify which one of the above
entries should be used when resolving a location.

Fixed: 436290252
Change-Id: I28f636bc33d80a4bca62dc50859d4d826a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1410834
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>