[fshost] report non-empty topo path for in process disk
The in-process disk we serve in recovery had an empty string path which
because of our logic around ramdisk children was tagging every other
device we found as a ramdisk. This messed with matching since ramdisks
tend to be treated special.

Bug: 458454009
Change-Id: I9211c759440845a4abe46da6816541cc695623a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416794
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>