[ffx] SSH key verifier command.
For eng and userdebug builds, downloads the authorized_keys from the
device via HTTP, and uses the result to verify whether or not matching
public keys for the device are stored locally. Searches the ssh agent,
the home directory, and the fuchsia directory (if set) for ssh keys.

Things to add:

-- Checking pub/private keys match.
-- Auto-fix issues with keys missing from home dir.

Bug: 436649104
Change-Id: Ie59055a22752bac844dd46c3d8bb0c966a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364587
Commit-Queue: Andrew Davies <awdavies@google.com>
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>