[fuchsia-merkle] Add MerkleRootBuilder and MerkleVerifier
MerkleRootBuilder is similar to MerkleTreeBuilder but it doesn't store
all of the interior hashes. All users of MerkleTreeBuilder except for
fxblob only care about the root so this is slightly more efficient for
that purpose. Fxblob also needs the leaf hashes which MerkleRootBuilder
is also able to generate.

BufferedMerkleRootBuilder is similar to MerkleRootBuilder but is able to
buffer data. This is useful when the client is receiving data that may
not always be block aligned.

MerkleVerifier uses the leaf hashes generated by MerkleRootBuilder to
verify the data later.

ReadSizedMerkleVerifier builds on MerkleVerifier and can save a
significant amount of memory when data is always verified multiple
blocks at a time. This is intended to be used in fxblob.

Mundane was replaced with bssl because bssl's Sha256 implements Send.

Bug: b/437178186
Change-Id: If85a49f8aecea2a1160edd2e9f967f92a6d54980
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371261
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>