[ramdevice_client] remove unneeded isolated devmgr dep
The ramdevice-client-test library had a group that included the driver
shim library, but it is not clear why it was there. The comment
explaining it didn't seem to be right. A couple of deps were missing
from sysconfig-client that it got transitively, which were just fixed up
directly.

The allowance for these to use the shim was also removed.

Change-Id: Icc8ef609850cca6546b2e3422cf90c59b1a5b4a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1387104
Reviewed-by: David Pursell <dpursell@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>