[microbenchmarks] Measure VMAR allocation
Benchmarks cost of VMAR allocation under a VMAR that is differently
loaded both in terms of total number of existing allocations, as well
density of existing allocations.

Presently ASLR performance is somewhat linear with respect to number
of existing allocations, and shows notable slowdowns at 10k. This
benchmark is added to track improvements to this.

Bug: 445988883
Change-Id: I8f48cb4191eba57fff02b2883b5d217efba7bffd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1412214
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>