[storage] Remove uses of FutureExt::boxed
From looking at the generated assembly for fxfs and the rust VFS, Rust
doesn't typically inline FutureExt::boxed calls. Without inlining, the
future is constructed on then stack, then the memory allocation is made,
and then the future is copied to the allocation. Replacing
FutureExt::boxed with Box::new/pin causes the memory allocation to be
made first and then the future is constructed directly on the heap.

Fixed: b/456517172
Change-Id: I1f4d639ff0ca6a6197c2bd33d893e02f4e580867
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405813
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>