[phys] Introduce Allocation::Extend and Allocation::Adopt
Allocation::Extend purpose is to provide a best effort tail extension for an
allocated range.

Allocation::Adopt purpose is to allow taking an existing range, and
manufacture an Allocation object to operate on the range.

These methods are a requirement for BootZbi operating on Allocation
instead of arbitrary ranges, thus also a blocker on BootZbi extending
ZBIs in place when possible.

Both now expose `Allocation::PageSize()` such that proper rounding can
be achieved.

Change-Id: I28e8ccfc1f22d474dc2d6c95218148230eb18e8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1293024
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>