[feedback] Use spontaneous_reboot_reason flag for crash signatures
Assembly is supplying a spontaneous_reboot_reason flag inside of a
feedback_config.json file. This CL parses the config and uses the value
to determine the crash signature for spontaneous reboots.

Note that "spontaneous" will temporarily be reported as "brief power
loss" until all products have had time to specify their desired reason.
This is because "spontaneous" is the default in assembly, but "brief
power loss" is the historical default in Feedback. Leaving "brief power
loss" as the default in Feedback will reduce signature churn.

Bug: 441569016
Test: `fx test //src/developer/forensics`
Test: spontaneous reboot @ 7acb62bad58c85df
Change-Id: I5b3b097cebb6ad719a41b73030967f2cf2ac4530
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1419135
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Commit-Queue: Thomas Gales <tgales@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>