[fxtest][gemini] Highlight key stack trace lines in Gemini Analysis
This change makes it so that the key lines of the stack trace are
highlighted in color in the stack trace, instead of being listed
underneath. The lines outputted in red is deemed the most key line to
solving the issue, and the lines outputted in yellow are lines that
could potentially be used to debug the issue.

Test: Added unit tests to gemini_analysis_test to verify the new output functionality.
Change-Id: Ieca48025b819fdb10f6f39deaadb88595119c98f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416955
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Hardik Singh <singhhardik@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>