[phys] BootZbi::Load boot tests
Add boot tests that exercise the BootZbi::Load and Init path.
The cases under consideration involved whether the kernel can or not be
loaded in place, and whether the Data ZBI fits in the allocation or not,
and whether it can be extended in place.

The kernel dimension is handled by generating a ZBI that is kernel
aligned (it can be loaded in place) and conditionally inserting a 1 byte
long discard item before it(cannot be loaded in place).

The DataZbi checks are performed based on the initial allocation
dimension, and the extendendability can be guaranteed by overallocating
and then shrinking the allocation such that there are N bytes available
at the tail. And it can be guaranteed that it wont be extendable in
place(and force relocation) by allocating the tail when shrinking the
allocation, hence no bytes are available after it.

Change-Id: Ib44bf86ba7ad965ab9191a75aab6294e576f44b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1333918
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>