[starnix] "Sniff" a longer process name from argv[0].
Linux's uapi is limited to a comm value of 16 null-terminated
bytes, which is not long enough to differentiate between many
processes when the comm is set naively from the binary's
basename.

We use the comm value for Zircon process names because it's the
only logical task name in the uapi that always involves a
syscall. argv[0]'s basename is often the most useful name for a
process, but the most useful names in systems like ART get
written to argv[0] well after the task has started and don't
involve a syscall.

Using the comm value for Zircon process names makes it harder for
Fuchsia-side tools to understand which process is which if e.g.
there are a number of processes whose binaries share a common
prefix. This is especially challenging when Starnix's own config
relies on process names for things like rt_role_overrides.

This change attempts to improve the situation for all of the
above challenges by sniffing a longer comm for internal usage
when argv[0] is a superstring of the user-provided comm value.
In practice, this seems to get us better Zircon names for many
tasks in our current system without risking clobbering names
that are intentionally provided by userspace.

Bug: 455850151
Bug: 321273720

Change-Id: Ic5067953314cb4b1dea99ee307e7619530431415
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1406011
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Adam Perry <adamperry@google.com>