[starnix][rcu] Add RcuCache
RcuCache is a cache that uses RCU to provide thread-safe access to a
hash map. If the number of entries in the cache exceeds its capacity,
entries are evicted in FIFO order.

In the future, we might elaborate this data structure to retain recently
accessed items, for example using the CLOCK algorithm.

Bug: b/445230228
Change-Id: Ia498c9bd5c1676455cfa38e98b6a542ee6eafb6b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420035
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>