[zbi] Allow for providing an embedded ZBI
Allow for running the ZBI tool against a potentially embedded ZBI(in
some other file). The check is less strict about the size of the file
being exactly the length described in the header, and just check what
the header states to the contents length. This allows in the case of a
ZBI with an appended BOOTCONFIG to be checked, so that their contents
can be verified.

BUG: 408051286
Change-Id: I5c29be0e4ff9b9f3bdafcbce99ed81f8a2e011ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1417074
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>