[kernel][scheduler] Make Rescheduling with arch_blocking_disallowed an error
Prior to this change, Scheduler::Reschedule and
Scheduler::RescheduleMask and would pend a preemption when called in
an arch_blocking_disallowed context.  That is, they pend a preemption
and return rather than continue down into RescheduleCommon.  This
change removes the special case logic for arch_blocking_disallowed.

Now, calling Reschedule or RescheduleMask with
arch_blocking_disallowed is an error and will result in the existing
RescheduleCommon assert firing.

The motivation for this change is to remove the only remaining
non-assert use of arch_blocking_disallowed and to simplify the
reschedule logic in preparation for use of mp_reschedule_self.

Bug: 42143537
Bug: 457759524
Change-Id: I065e3739030151bdfe236f2dc77becf269d7ffef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1415586
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>